DB Overhaul
I've had to overhaul the db setup as it was stuffing us around with the
creation of different tables.  The way I had thought would be best was
to keep the table creation in different classes but that was a bad idea.

All DB work is now done through the DatabaseHelper class.  I've also
moved FoodItem to its own class.

Sorry for the fuck around :(