UserTraits Table Functionality Added
I've added the required functionality for the UserTraits table which
should allow Taylor to complete settings functionality.

The adding to the txt file is tricky though and I've tried to explain it
as best as possible in the code.  Please ask if you have questions.

This new functionality is all untested and may require work, although I
think I've got all the possible issues covered.