More DB additions
I've gone through and added ConsumedFood Table, and OrderHeader table.
I've decided not to implement LookupOrderType but rather use just a
series of int's to look after that.  I'll make them global eventually.

Not all the queries are in and working yet and tbh I haven't even tested
the work that I've done tonight.  I'll look at adding more queries and
refining the ones that are there tomorrow night.