FUCK YOU SEARCH, FUCK YOU
So as the title alludes to, I had heaps of trouble trying to implement
this search function.  Currently as it stands, you click 'Add Entry' and
it displays a search bar at the top.  You can click in here and start
typing and results appear, and how we want
them!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

Now we need to add them to a nice list and eventually save that to the
db bahahahah GFL.

Any whooooo, there are many unused layout files, a couple of classes etc
that we won't use but we'll leave them for now until we are close to
release.

Don't ask me exactly how I got it to work, it just does lolz.