Clicking on Search option now adds item to list
I've implemented a ExpandableList to the Add Entry activity.  This
basically takes the item the user clicks on after searching and adds it
to a List.  This list will eventually contain an image of the item, the
item name and if the user clicks the item it will drop down and show the
nutrition information.

I've done a fresh install of the app after fixing this and have not
touched the code since so it should work fine :)